Add an updated autolog script
Previous version does not run with a not deprecated interpreter and its
source is not clear. This replacement is part of Gnulib (see the header
comments).

There are also modified versions of the removed scripted that may
work (which have a timestamp of 2 month ago instead of 17 years), such
as
https://github.com/envmodules/modules/blob/a6249f3bed42d220061ceb15acad0075ffc36c43/script/gitlog2changelog.py.in
which can be studied in the future.  May this one be used with a
junction of hooks and some handwritten comments of the recent history of
the project in the future commits.

The new script is fairy more complex than the previous one but its a
commonly accepted script for such a purpose (see GNU Coding Standards
for more information).