Remove Python wrapper
The Python version is basically not supported at this point (2) and
having this adds a burden to developers and maintainers both alike.

This can be added back later when the project is in a better state
either directly or by a dependency or in another project altogether.
Personally, it would be better maintained if added externally with other
languages like Rust next to it.