Add a minimalistic linter action
This action, triggered on push or pull, fails if clang-format fails.
This addresses the comment referenced by [1] on #28 (formatting all the
sources following the minimal path examples on #23).

Note that the "sources" directory is hardcoded on the action file
pushed. Assuming the change of source directory is a fundamental change
with a lot of considerations, no documentations were added to mention
that the value used in this file must be synchronized with the directory
name on the tree. Suppose this comment suffices.

[1]: https://github.com/persiancal/jcal/pull/28#issuecomment-2823474321