Remove extra def and cpy of jtm structs
Some internal (`in_*`) functions define an inner value and copy it over
a given buffer. This complicates lifetimes and an extra copy is
introduced which can be prevented.

In this minor change, these functions get their mutable pointers and
directly assign on them instead of allocating a new instance and copying
it then freeing the original.